feat: 로그인 후 추적 요청 받도록 수정
